choer: rename config.js
