+++ lib/compile.config.js
@@ -1,8 +1,7 @@
 var webpack = require('webpack');
 var ExtractTextPlugin = require('extract-text-webpack-plugin');
-var getBabel = require('./lib/getBabel');
-var getTs = require('./lib/getTs');
-
+var getBabel = require('./getBabel');
+var getTs = require('./getTs');
 
 var babelLoader = require.resolve('babel-loader');
 var cssLoader = require.resolve('css-loader');
