+++ lib/compiler.js
@@ -9,7 +9,7 @@ var memoryFs = new MemoryFileSystem();
 var merge = require('lodash.merge');
 var gutil = require('gulp-util');
 
-var defaultConfig = require('../webpack.config.js');
+var defaultConfig = require('./compile.config.js');
 var tempfilePath;
   /**
    * 异步 webpack 编译代码
@@ -24,7 +24,9 @@ module.exports = function(code, config, callback) {
     var tempfileRelativePath = path.resolve(config.context, tempfilePath);
 
     config = merge({}, defaultConfig, config, {
-      entry: tempfileRelativePath
+      entry: {
+        common: tempfileRelativePath
+      }
     });
 
     fs.writeFileSync(tempfileRelativePath, code);
@@ -60,12 +62,12 @@ module.exports = function(code, config, callback) {
       console.log(stats.toString({
         colors: true
       }));
-      return callback('webpack compile error')
+      return callback('webpack compile error');
     }
   });
 
   compiler.plugin('failed', function(err) {
-    console.log(err)
+    console.log(err);
   });
 
   compiler.plugin('after-emit', function(compilation, callback) {
